component fixes
