+++ resources/views/layouts/footer.blade.php
@@ -68,6 +68,4 @@
             </div>
         </div>
     </div>
-
-
 </footer>
